+++ src/Reflex/Vty/Widget/Layout.hs
@@ -6,7 +6,6 @@ Description: Monad transformer and tools for arranging widgets and building scre
 
 module Reflex.Vty.Widget.Layout where
 
-import Control.Applicative (liftA2)
 import Control.Monad.Catch (MonadCatch, MonadThrow, MonadMask)
 import Control.Monad.Morph
 import Control.Monad.NodeId (MonadNodeId(..), NodeId)
@@ -143,6 +142,8 @@ instance (HasImageWriter t m, MonadFix m) => HasImageWriter t (Focus t m) where
 instance (HasFocusReader t m, Monad m) => HasFocusReader t (Focus t m)
 
 instance (HasTheme t m, Monad m) => HasTheme t (Focus t m)
+
+instance (HasColorProfile t m, Monad m) => HasColorProfile t (Focus t m)
 
 instance (Reflex t, MonadFix m, MonadNodeId m) => HasFocus t (Focus t m) where
   makeFocus = do
@@ -450,6 +451,8 @@ instance (HasDisplayRegion t m, HasImageWriter t m, MonadFix m) => HasImageWrite
 instance (HasFocusReader t m, Monad m) => HasFocusReader t (Layout t m)
 
 instance (HasTheme t m, Monad m) => HasTheme t (Layout t m)
+
+instance (HasColorProfile t m, Monad m) => HasColorProfile t (Layout t m)
 
 instance (Monad m, MonadNodeId m, Reflex t, MonadFix m) => HasLayout t (Layout t m) where
   axis o c (Layout x) = Layout $ do